requirements.txt
Flask - installed
Flask-Session - installed
psycopg2-binary - installed
SQLAlchemy - installed
